Email validation, ResponseStatusException on Read and Delete

+++ src/main/java/com/project/projetdatauser/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.project.projetdatauser.controller;
 
-import com.project.projetdatauser.exceptions.UserNotFoundException;
 import com.project.projetdatauser.model.User;
 import com.project.projetdatauser.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +8,7 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.Optional;
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -19,34 +18,26 @@ public class UserController {
     private UserRepository userRepository;
 
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public void create(@RequestBody User user) {
+    public void create(@Valid @RequestBody User user) {
         userRepository.save(user);
     }
 
     @GetMapping(value = "/{id}")
-    public User read(@PathVariable String id) {
-
-        Optional<User> user = userRepository.findById(id);
-        if(!user.isPresent()){
-            throw new UserNotFoundException("id- " + id);
-        }
-        return user.get();
-        /*
-        try {
-            return userRepository.findOneById(id);
-        } catch (UserNotFoundException ex){
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "User Not Found", ex);
-        }*/
+    public User read(@PathVariable String id)  {
+        User usr = userRepository.findOneById(id);
+        if (usr == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found To Read");
+        return usr;
     }
 
     @RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public void update(@RequestBody User user) {
+    public void update(@Valid @RequestBody User user) {
         userRepository.save(user);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public void delete(@PathVariable String id) {
+        User usr = userRepository.findOneById(id);
+        if (usr == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not Found To Delete");
         userRepository.deleteById(id);
     }
 }
